pyproject.toml
[project]
name = "high-five-recognition-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
